+++ lib/src/widgets/radio/ds_radio_tile.widget.dart
@@ -17,18 +17,22 @@ class DSRadioTile<T> extends StatelessWidget {
     required this.onChanged,
     required this.title,
     required this.groupValue,
+    this.subTitle,
     this.isEnabled = true,
-    this.contentPadding,
+    this.contentPadding = const EdgeInsets.symmetric(vertical: 16.0),
   }) : super(key: key);
 
   /// Clickable radio group title allowing you to select an option.
-  final Widget? title;
+  final Widget title;
 
   /// Determines the grouping of buttons allowing one of them to be selected
   ///
   /// The <T> type determines which type of variable to use as [value],
   /// which can be int, String, or even an object.
   final T? groupValue;
+
+  /// Radio tile subtitle widget.
+  final Widget? subTitle;
 
   /// Allows you to enable or disable the use of the button group, setting a
   /// default value and preventing other options from being returned.
@@ -41,23 +45,41 @@ class DSRadioTile<T> extends StatelessWidget {
   final T value;
 
   /// Content spacing inner border size
-  final EdgeInsetsGeometry? contentPadding;
+  final EdgeInsetsGeometry contentPadding;
 
   @override
   Widget build(BuildContext context) {
-    return ListTile(
-      enabled: isEnabled,
-      dense: true,
-      horizontalTitleGap: 0,
-      onTap: () => onChanged!(value),
-      leading: DSRadio(
-        isEnabled: isEnabled,
-        value: value,
-        groupValue: groupValue,
-        onChanged: onChanged,
+    return InkWell(
+      onTap: isEnabled ? () => onChanged!(value) : null,
+      child: Padding(
+        padding: contentPadding,
+        child: Row(
+          children: [
+            DSRadio(
+              isEnabled: isEnabled,
+              value: value,
+              groupValue: groupValue,
+              onChanged: onChanged,
+              materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
+            ),
+            const SizedBox(
+              width: 8.0,
+            ),
+            Flexible(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  title,
+                  const SizedBox(
+                    height: 5.0,
+                  ),
+                  if (subTitle != null) subTitle!,
+                ],
+              ),
+            ),
+          ],
+        ),
       ),
-      title: title,
-      contentPadding: contentPadding,
     );
   }
 }