+++ lib/src/widgets/utils/ds_divider.widget.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+import '../../themes/colors/ds_colors.theme.dart';
+
+class DSDivider extends Divider {
+  const DSDivider({
+    super.key,
+    super.thickness = 1.0,
+    super.height = 1.0,
+    super.color = DSColors.neutralMediumWave,
+  });
+}